User logout and Fixing post deletion

+++ app/Http/Controllers/PostController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
 use App\Post;
 use Illuminate\Http\Request;
 
@@ -9,7 +10,7 @@ class PostController extends Controller
 {
     public function getDashboard()
     {
-        $post = Post::all();
+        $post = Post::OrderBy('created_at','desc')->get();
         return view('dashboard',['posts'=>$post]);
     }
 
@@ -30,10 +31,18 @@ class PostController extends Controller
     public function getDeletePost($post_id)
     {
         $post = Post::where('id',$post_id)->first();
+        if(Auth::user()!= $post->user){
+            return redirect()->back();
+        }
         /* This is also available 
            $post = find($post_id)->first(); 
         */
         $post->delete();
         return redirect()->route('dashboard')->with(['message'=>'Successfully Deleted!']);
+    }
+    public function getLogOut()
+    {
+        Auth::logout();
+        return redirect()->route('home');
     }
 }
